requirements.txt
pypdf
pdfplumber
pandas
openai
python-dotenv
